feat: add cache to playlistsService

+++ src/services/postgres/PlaylistsService.js
@@ -6,9 +6,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const ClientError = require('../../exceptions/ClientError');
 
 class PlaylistsService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   async addPlaylist({ name, owner }) {
@@ -24,21 +25,32 @@ class PlaylistsService {
     if (!result.rows[0].id) {
       throw new InvariantError('Playlist gagal ditambahkan');
     }
+    await this._cacheService.delete(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON playlists.id = collaborations.playlist_id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows;
+    try {
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON playlists.id = collaborations.playlist_id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
+      await this._cacheService.set(
+        `playlists:${owner}`,
+        JSON.stringify(result.rows),
+      );
+      // console.log(result.rows);
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -47,6 +59,9 @@ class PlaylistsService {
     if (!result.rows.length) {
       throw new NotFoundError('Playlist gagal dihapus. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 
   // Playlist_songs Start
@@ -61,6 +76,7 @@ class PlaylistsService {
     if (!result.rows[0].id) {
       throw new InvariantError('Lagu gagal ditambahkan ke playlist');
     }
+    await this._cacheService.delete(`songs_playlist:${songId}`);
     return result.rows[0].id;
   }
 
@@ -71,21 +87,34 @@ class PlaylistsService {
       values: [id, playlistId, songId, credentialId, action, time],
     };
     await this._pool.query(query);
+
+    await this._cacheService.delete(`playlist_activities:${playlistId}`);
   }
 
   async getPlaylistByPlaylistsId(playlistId) {
-    const query = {
-      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists INNER JOIN users ON users.id = playlists.owner WHERE playlists.id = $1',
-      values: [playlistId],
-    };
-
-    const result = await this._pool.query(query);
-    if (!result.rows) {
-      throw new NotFoundError(
-        'Anda tidak memiliki playlist by getPlaylistById',
+    try {
+      const result = await this._cacheService.get(
+        `playlist_activities:${playlistId}`,
       );
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: 'SELECT playlists.id, playlists.name, users.username FROM playlists INNER JOIN users ON users.id = playlists.owner WHERE playlists.id = $1',
+        values: [playlistId],
+      };
+
+      const result = await this._pool.query(query);
+      if (!result.rows) {
+        throw new NotFoundError(
+          'Anda tidak memiliki playlist by getPlaylistById',
+        );
+      }
+      await this._cacheService.set(
+        `playlist_activities:${playlistId}`,
+        JSON.stringify(result.rows[0]),
+      );
+      return result.rows[0];
     }
-    return result.rows[0];
   }
 
   async getPlaylistWithSong(playlistId) {
@@ -99,6 +128,7 @@ class PlaylistsService {
         'Anda tidak memiliki playlist by getPlaylistWithSong',
       );
     }
+
     return result.rows;
   }
 
@@ -113,21 +143,30 @@ class PlaylistsService {
         'Lagu gagal dihapus dari playlist, song id tidak ditemukan',
       );
     }
+    const { id } = result.rows[0].id;
+    await this._cacheService.delete(`playlist_activities:${id}`);
   }
 
   async getPlaylistActivitiesByPlaylistsId(playlistId) {
-    const query = {
-      text: 'SELECT users.username, songs.title, playlist_song_activities.action, playlist_song_activities.time FROM playlist_song_activities LEFT JOIN users ON playlist_song_activities.user_id = users.id LEFT JOIN songs ON playlist_song_activities.song_id = songs.id WHERE playlist_id = $1',
-      values: [playlistId],
-    };
-
-    const result = await this._pool.query(query);
-    if (!result.rows) {
-      throw new NotFoundError(
-        'Anda tidak memiliki playlist by getPlaylistById',
+    try {
+      const result = await this._cacheService.get(
+        `playlist_activities:${playlistId}`,
       );
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: 'SELECT users.username, songs.title, playlist_song_activities.action, playlist_song_activities.time FROM playlist_song_activities LEFT JOIN users ON playlist_song_activities.user_id = users.id LEFT JOIN songs ON playlist_song_activities.song_id = songs.id WHERE playlist_id = $1',
+        values: [playlistId],
+      };
+
+      const result = await this._pool.query(query);
+      if (!result.rows) {
+        throw new NotFoundError(
+          'Anda tidak memiliki playlist by getPlaylistById',
+        );
+      }
+      return result.rows;
     }
-    return result.rows;
   }
 
   async checkSongOnList(playlistId, songId) {
